refactor: extract pure game engine module
