+++ index.html
@@ -38,6 +38,7 @@
 
   <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.6.4/jquery.min.js"></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/chessboard.js/1.0.0/chessboard-1.0.0.min.js"></script>
+  <script src="src/gameEngine.js"></script>
   <script src="app.js"></script>
 </body>
 </html>
